embedding-and-rag/pyproject.toml
[project]
name = "embedding-and-rag"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.13",
    "langchain-text-splitters>=1.1.1",
    "pinecone>=8.1.0",
    "qdrant-client>=1.17.1",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
]
